+++ syndicate/connection/kinesis_connection.py
@@ -69,5 +69,9 @@ class KinesisConnection(object):
                 result.extend(response['StreamNames'])
         return result
 
-    def remove_stream(self, stream_name):
+    def remove_stream(self, stream_name, log_not_found_error=True):
+        """
+        log_not_found_error parameter is needed for proper log handling in the
+        retry decorator
+        """
         self.client.delete_stream(StreamName=stream_name)