+++ syndicate/core/resources/ebs_resource.py
@@ -46,7 +46,7 @@ class EbsResource(BaseResource):
         if not response:
             response = self.ebs_conn.describe_applications([name])
         if not response:
-            return
+            return {}
         return {
             arn: build_description_obj(response, name, meta)
         }
@@ -225,16 +225,18 @@ class EbsResource(BaseResource):
         })
 
     def remove_ebs_apps(self, args):
-        self.create_pool(self._remove_ebs_app, args)
+        return self.create_pool(self._remove_ebs_app, args)
 
     @unpack_kwargs
     def _remove_ebs_app(self, arn, config):
         app_name = config['resource_name']
         try:
-            self.ebs_conn.remove_app(app_name)
+            self.ebs_conn.remove_app(app_name, log_not_found_error=False)
             _LOG.info(f'EBS app {app_name} was removed.')
+            return {arn: config}
         except ClientError as e:
             if e.response['Error']['Code'] == 'ResourceNotFoundException':
                 _LOG.warn(f'EBS app {app_name} is not found')
+                return {arn: config}
             else:
                 raise e